Add extracting table context to own folder
