+++ packages/common/src/ui/layout/tables/components/DataRow/DataRow.tsx
@@ -3,7 +3,7 @@ import TableRow from '@mui/material/TableRow';
 import TableCell from '@mui/material/TableCell';
 import { Column } from '../../columns/types';
 import { DomainObject } from '@common/types';
-import { useExpanded, useDisabled, useRowStyle } from '../../context';
+import { useExpanded, useIsDisabled, useRowStyle } from '../../context';
 import { Collapse, Fade } from '@mui/material';
 
 interface DataRowProps<T extends DomainObject> {
@@ -29,7 +29,7 @@ export const DataRow = <T extends DomainObject>({
 }: DataRowProps<T>): JSX.Element => {
   const hasOnClick = !!onClick;
   const { isExpanded } = useExpanded(rowData.id);
-  const { isDisabled } = useDisabled(rowData.id);
+  const { isDisabled } = useIsDisabled(rowData.id);
   const { rowStyle } = useRowStyle(rowData.id);
 
   const onRowClick = () => onClick && onClick(rowData);
