+++ packages/common/src/ui/layout/tables/context/TableContext.ts
@@ -1,6 +1,3 @@
-import isEqual from 'lodash/isEqual';
-import { useLocalStorage, GroupByItem } from './../../../../localStorage';
-import { useCallback, useEffect } from 'react';
 import create, { UseStore } from 'zustand';
 import createContext from 'zustand/context';
 import { AppSxProp } from '../../../../styles';
@@ -259,116 +256,3 @@ export const createTableStore = (): UseStore<TableStore> =>
       });
     },
   }));
-
-interface UseExpandedControl {
-  isExpanded: boolean;
-  toggleExpanded: () => void;
-}
-
-export const useExpanded = (rowId: string): UseExpandedControl => {
-  const selector = useCallback(
-    (state: TableStore) => {
-      return {
-        rowId,
-        isExpanded: state.rowState[rowId]?.isExpanded ?? false,
-        toggleExpanded: () => state.toggleExpanded(rowId),
-      };
-    },
-    [rowId]
-  );
-
-  const equalityFn = (
-    oldState: ReturnType<typeof selector>,
-    newState: ReturnType<typeof selector>
-  ) =>
-    oldState?.isExpanded === newState?.isExpanded &&
-    oldState.rowId === newState.rowId;
-
-  return useTableStore(selector, equalityFn);
-};
-
-interface UseDisabledControl {
-  isDisabled: boolean;
-  // toggleDisabled: () => void; // TODO: Is this needed?
-}
-
-export const useDisabled = (rowId: string): UseDisabledControl => {
-  const selector = useCallback(
-    (state: TableStore) => {
-      return {
-        rowId,
-        isDisabled: state.rowState[rowId]?.isDisabled ?? false,
-      };
-    },
-    [rowId]
-  );
-
-  const equalityFn = (
-    oldState: ReturnType<typeof selector>,
-    newState: ReturnType<typeof selector>
-  ) =>
-    oldState?.isDisabled === newState?.isDisabled &&
-    oldState.rowId === newState.rowId;
-
-  return useTableStore(selector, equalityFn);
-};
-
-export interface UseRowStyleControl {
-  rowId?: string;
-  rowStyle: AppSxProp;
-  setRowStyle: (id: string, rowStyle: AppSxProp) => void;
-  setRowStyles: (ids: string[], rowStyle: AppSxProp) => void;
-}
-
-export const useRowStyle = (rowId?: string): UseRowStyleControl => {
-  const selector = useCallback(
-    (state: TableStore) => {
-      return {
-        rowId,
-        rowStyle: rowId ? state.rowState[rowId]?.style ?? {} : {},
-        setRowStyle: (id: string, style: AppSxProp) =>
-          state.setRowStyle(id, style),
-        setRowStyles: (ids: string[], style: AppSxProp) =>
-          state.setRowStyles(ids, style),
-      };
-    },
-    [rowId]
-  );
-
-  const equalityFn = (
-    oldState: ReturnType<typeof selector>,
-    newState: ReturnType<typeof selector>
-  ) => isEqual(oldState.rowStyle, newState.rowStyle);
-
-  return useTableStore(selector, equalityFn);
-};
-
-interface IsGroupedControl {
-  isGrouped: boolean;
-  toggleIsGrouped: () => void;
-}
-
-export const useIsGrouped = (key: keyof GroupByItem): IsGroupedControl => {
-  const selector = useCallback(({ setIsGrouped }: TableStore) => {
-    return { setIsGrouped };
-  }, []);
-  const { setIsGrouped } = useTableStore(selector);
-  const [groupByItem, setGroupByItem] = useLocalStorage('/groupbyitem', {
-    outboundShipment: false,
-    inboundShipment: false,
-  });
-
-  const toggleIsGrouped = useCallback(() => {
-    const newVal = !groupByItem?.[key];
-    setGroupByItem({ ...groupByItem, [key]: newVal });
-  }, [groupByItem, key, setGroupByItem, setIsGrouped]);
-
-  useEffect(() => {
-    // Sync the table state up with the local storage state.
-    // Syncing the states rather than explicitly setting in the callback
-    // as we need to do this on the initial mount regardless.
-    setIsGrouped(!!groupByItem?.[key]);
-  }, [groupByItem?.[key]]);
-
-  return { isGrouped: !!groupByItem?.[key], toggleIsGrouped };
-};
