+++ packages/common/src/ui/layout/tables/context/index.ts
@@ -1,1 +1,2 @@
 export * from './TableContext';
+export * from './hooks';
